pyproject.toml
[project]
name = "chimera"
version = "0.1.0"
description = "CHIMERA: Autonomous News-Driven Polymarket Trading Agent"
requires-python = ">=3.11"
dependencies = [
    # Core
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "httpx>=0.27.0",

    # Polymarket
    "py-clob-client>=0.34.0",

    # LLM Backends
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "google-generativeai>=0.8.0",

    # NLP / ML
    "sentence-transformers>=3.0.0",
    "hdbscan>=0.8.38",
    "datasketch>=1.6.0",
    "numpy>=1.26.0",

    # Knowledge Graph
    "networkx>=3.3",

    # Data Ingestion
    "feedparser>=6.0.0",
    "beautifulsoup4>=4.12.0",

    # Utils
    "python-dotenv>=1.0.0",
    "structlog>=24.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.pytest.ini_options]
asyncio_mode = "auto"
